Finish fleshing out packet buffers
This design is a little more obtuse, but the goal is to keep each individual
struct simple and focused. I removed the object pool and may add it back in
later if it seems necessary. From what I've read, jemalloc does a pretty good
job of allocating buffers of the size we need (it's considered a "small"
allocation) so it may be fine to just not use an object pool at all.